+++ client/tests/unit/mixins/decision-owner-test.js
@@ -1,0 +1,72 @@
+import DecisionOwnerMixin from 'tahi/mixins/decision-owner';
+import Ember from 'ember';
+import { make } from 'ember-data-factory-guy';
+import { test } from 'qunit';
+import moduleForAcceptance from 'tahi/tests/helpers/module-for-acceptance';
+
+const DecisionOwnerObject = Ember.Object.extend(DecisionOwnerMixin);
+
+moduleForAcceptance('Unit | Mixin | decision owner');
+
+test('decisionsAscending with many decisions, including a draft decision', function(assert) {
+  const decisionOwner = DecisionOwnerObject.create({
+    decisions: [
+      make('decision', { majorVersion: 1, minorVersion: 0, draft: false}),
+      make('decision', { majorVersion: 0, minorVersion: 0, draft: false}),
+      make('decision', { majorVersion: 0, minorVersion: 1, draft: false}),
+      make('decision', { majorVersion: null, minorVersion: null, draft: true}),
+      make('decision', { majorVersion: 1, minorVersion: 1, draft: false})
+    ]
+  });
+  const versions = decisionOwner.get('decisionsAscending').map(function(decision) {
+    return decision.getProperties('minorVersion', 'majorVersion', 'draft');
+  });
+  const expectedVersions = [
+    { majorVersion: 0, minorVersion: 0, draft: false},
+    { majorVersion: 0, minorVersion: 1, draft: false},
+    { majorVersion: 1, minorVersion: 0, draft: false},
+    { majorVersion: 1, minorVersion: 1, draft: false},
+    { majorVersion: null, minorVersion: null, draft: true}
+  ];
+  assert.deepEqual(versions, expectedVersions);
+});
+
+test('latestDecision when there are two decisions', function(assert) {
+  const decisionOwner = DecisionOwnerObject.create({
+    decisions: [
+      make('decision', { majorVersion: 1, minorVersion: 0, draft: false}),
+      make('decision', { majorVersion: 0, minorVersion: 0, draft: false})
+    ]
+  });
+  const version = decisionOwner.get('latestDecision').getProperties('majorVersion', 'minorVersion', 'draft');
+  const expectedVersion = { majorVersion: 1, minorVersion: 0, draft: false };
+  assert.deepEqual(version, expectedVersion);
+});
+
+test('latestDecision when there are zero decisions', function(assert) {
+  const decisionOwner = DecisionOwnerObject.create({
+    decisions: []
+  });
+  assert.equal(decisionOwner.get('latestDecision'), null);
+});
+
+test('latestDecision when version numbers change', function(assert) {
+  const decisionOwner = DecisionOwnerObject.create({
+    decisions: [
+      make('decision', {majorVersion: 1, minorVersion: 0, draft: false}),
+      make('decision', {majorVersion: 0, minorVersion: 0, draft: false})
+    ]
+  });
+  const version = decisionOwner.get('latestDecision').getProperties('majorVersion', 'minorVersion', 'draft');
+  const expectedVersion = { majorVersion: 1, minorVersion: 0, draft: false };
+  assert.deepEqual(version, expectedVersion);
+
+  // now change the majorVersion of the 0.0 decision so it becomes the latest
+  Ember.run( () => {
+    decisionOwner.get('decisions.lastObject').set('majorVersion', 2);
+  });
+  const mutatedVersion = decisionOwner.get('latestDecision')
+    .getProperties('majorVersion', 'minorVersion', 'draft');
+  const expectedMutatedVersion = { majorVersion: 2, minorVersion: 0, draft: false };
+  assert.deepEqual(mutatedVersion, expectedMutatedVersion);
+});