+++ engines/tahi_standard_tasks/client/app/models/reviewer-report-task.js
@@ -1,15 +1,9 @@
-import Ember from 'ember';
+import DecisionOwner from 'tahi/mixins/decision-owner';
 import DS from 'ember-data';
 import Task from 'tahi/models/task';
 
-export default Task.extend({
+export default Task.extend(DecisionOwner, {
   paperReview: DS.belongsTo('paperReview'),
   decisions: DS.hasMany('decision'),
-  isSubmitted: DS.attr('boolean'),
-
-  previousDecisions: Ember.computed.alias('task.paper.previousDecisions'),
-
-  decision: Ember.computed('decisions', function() {
-    return this.get('decisions').findBy('draft', true);
-  })
+  isSubmitted: DS.attr('boolean')
 });